% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{arch_to_assignment}
\alias{arch_to_assignment}
\title{Convert Architecture(s) back to Assignment Matrices}
\usage{
arch_to_assignment(arch, .id = NULL)
}
\arguments{
\item{arch}{data.frame of architecture(s) form}

\item{.id}{name of architecture unique id column. (Typically is \code{"id"}.) Default is NULL.}
}
\description{
Intended to be used on data.frames generated by \code{enumerate_assignment()}.
}
\examples{
a = tribble(
  ~id, ~d1_a1_b1, ~d1_a2_b1, ~d1_a2_b2, ~d1_a2_b3,
  1,    1,        0,         1,         0,
  2,    1,        0,         0,         1
)
arch_to_assignment(arch = a, .id = "id")
}
\author{
Tim Fraser
}
