% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{count_ds}
\alias{count_ds}
\title{Count N for Downselecting}
\usage{
count_ds(n, k)
}
\arguments{
\item{n}{total number of items available to choose from}

\item{k}{number of items to choose}
}
\description{
Function to count total number of ways you can select \code{k} items from \code{n} items.
}
\note{
count_ds(n = 10, k = 2)
}
\author{
Tim Fraser
}
