% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{dominates}
\alias{dominates}
\title{Domination Checker}
\source{
Nozomi Hitomi, 11/17/2015
This function returns -1 if s1 dominates s2, 1 if s2 dominates
s1, 0 if neither dominates the other. Each solution should be a vector
of objectives. Assumes minimization in all objectives
}
\usage{
dominates(s1, s2)
}
\arguments{
\item{s1}{first matrix of metrics}

\item{s2}{second matrix of metrics}
}
\description{
Domination Checker
}
\author{
Tim Fraser (adapted from Nozomi Hitomi)
}
