% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate_binary}
\alias{enumerate_binary}
\title{Enumerate Binary Decisions into Architectures}
\usage{
enumerate_binary(n = 2, forbidden = NULL, .did = 1, .id = FALSE)
}
\arguments{
\item{n}{number of decisions. Must be at least 1.}

\item{forbidden}{...}

\item{.did}{(integer, optional) unique ID of first decision. Used to make column names for sequential decisions. Defaults to 1.}

\item{.id}{(logical, optional) Whether or not to add a unique ID for each row. Defaults to FALSE.}
}
\description{
Function for enumerating \code{n} binary decisions into a \code{data.frame} of encoded architectures.
}
\examples{
enumerate_binary(n = 2)
enumerate_binary(n = 2, forbidden = c(NA, 1))
enumerate_binary(n = 4, .did = 3)

}
\author{
Tim Fraser
}
