% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate_partition}
\alias{enumerate_partition}
\title{Enumerate Partitions into Architectures}
\usage{
enumerate_partition(
  n,
  k,
  min_times = NULL,
  max_times = NULL,
  .did = 1,
  .id = FALSE
)
}
\arguments{
\item{n}{total number of items to be partitioned}

\item{k}{total number of partitions to be made (categories to assign)}

\item{min_times}{(integer) a k-length vector of minimum number of times each respective category must be assigned.}

\item{max_times}{(integer) a k-length vector of maximum number of times each respective category must be assigned.}

\item{.did}{unique decision id.}

\item{.id}{(logical) Include a unique ID for each row? Defaults to FALSE.}
}
\description{
Enumerate Partitions into Architectures
}
\examples{
enumerate_partition(n = 4, k = 2)
enumerate_partition(n = 4, k = 2, min_times = c(1,3))
enumerate_partition(n = 4, k = 2, min_times = c(2,2))
enumerate_partition(n = 4, k = 2, max_times = c(2,2))
enumerate_partition(n = 4, k = 2, max_times = c(2,2), .did = 4)
}
\author{
Tim Fraser
}
