% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_coords}
\alias{get_coords}
\title{Get Coordinates for Network Visual from igraph object}
\usage{
get_coords(g, layout = "kk", vars = "name")
}
\arguments{
\item{g}{igraph graph object}

\item{layout}{ggraph-style character string layout option. Default is 'kk' for Kamada-Kawai layout}

\item{vars}{character vector of names of node traits from the igraph object to return}
}
\description{
Returns a list object include \code{nodes} and \code{edges}, which can be passed to \code{geom_point()} and \code{geom_curve()} respectively.
}
