% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_global}
\alias{get_global}
\title{Get Global Measures for an AHP table}
\usage{
get_global(data, format = "wide")
}
\arguments{
\item{data}{data.frame output from \code{get_priority()}, containing both alternatives and criteria.}

\item{format}{"wide" or "long". Describes output format of data.frame. Defaults to \code{"wide"}.}
}
\description{
Get Global Measures for an AHP table
}
\author{
Tim Fraser
}
