% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes from edgelist or adjacency matrix}
\usage{
get_nodes(edges = NULL, a = NULL)
}
\arguments{
\item{edges}{Optionally, an edgelist where the first column is \code{from} vertices and the second column is \code{to} vertices.}

\item{a}{Optionally, an adjacency matrix}
}
\description{
PART 1 FUNCTIONS
This script contains a series of functions that allow for stakeholder network analysis and supporting actions.
}
\details{
Some of these are internal functions, while others are functions intended to be used by actual R users.
If you encounter an issue with any of these functions, please tag it in a Github Issue.
}
\author{
Tim Fraser
}
