% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{get_percentages}
\alias{get_percentages}
\title{Tally percentages of each alternative for a decision variable in an architecture.}
\usage{
get_percentages(x, did = "d2")
}
\arguments{
\item{x}{vector of alternatives for a specific decision. As long as there are rows in the architectural matrix.}

\item{did}{string label of decision variable, eg. "d2" for decision 2}
}
\description{
Tally percentages of each alternative for a decision variable in an architecture.
}
