% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_priority}
\alias{get_priority}
\title{Get Priority Measure for an AHP table}
\usage{
get_priority(data, type = "criteria")
}
\arguments{
\item{data}{data.frame of pairwise comparisons of alternatives or criteria, formatted like \code{archr::pairs_alts} or \code{archr::pairs_criteria}}

\item{type}{"criteria" and/or "alt". Describes what kind of pairwise comparisons are provided in \code{data}. Use \code{c("criteria", "alt")} to get both.}
}
\description{
Get Priority Measure for an AHP table
}
\author{
Tim Fraser, PhD
}
