% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheet.R
\name{get_sheet}
\alias{get_sheet}
\title{Get Google Sheet Download Link}
\usage{
get_sheet(url = NULL, docid = NULL, gid = 0)
}
\arguments{
\item{url}{view-allowed Google Sheets share link (Optional. Must supply either \code{url} or \code{docid})}

\item{docid}{ID for that overall Google document. (Optional. Must supply either \code{url} or \code{docid})}

\item{gid}{sheet ID for that specific sheet (1 google sheets contains multiple sheets)}
}
\description{
Let's write a quick function for grabbing a specific google sheet using its share link.
}
