% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_entropy.R
\name{ig}
\alias{ig}
\title{Information Gain for Feature}
\usage{
ig(good, feature = NULL, .all = FALSE)
}
\arguments{
\item{.all}{binary TRUE/FALSE value, indicating whether to return a list of results, or just the stat data.frame}

\item{.feature}{a binary TRUE/FALSE vector indicating whether that architecture displays a specific feature.}

\item{.good}{a binary TRUE/FALSE vector indicating whether that architecture is considered 'good'}
}
\description{
Information Gain for Feature
}
\author{
Tim Fraser
}
