% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{me}
\alias{me}
\title{Calculate main effect}
\usage{
me(data, decision = "d2", value = 1, metric = "m1")
}
\arguments{
\item{data}{data.frame}

\item{decision}{string: name of decision variable i}

\item{value}{integer: alternative value for decision variable i}

\item{metric}{string: name of performance metric under evaluation}
}
\description{
Calculate main effect
}
