% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{sensitivity}
\alias{sensitivity}
\title{Get sensitivity score}
\usage{
sensitivity(data, decision_i = "d2", metric = "m1")
}
\arguments{
\item{data}{data.frame}

\item{decision_i}{string: name of decision variable i}

\item{metric}{string: name of performance metric under evaluation}
}
\description{
Get sensitivity score
}
